split actions into their own component
